Week-X Refactor Flask Routes
