+++ backend-flask/lib/helpers.py
@@ -1,0 +1,7 @@
+
+
+def model_json(model):
+  if model['errors'] is not None:
+    return model['errors'], 422
+  else:
+    return model['data'], 200
